requirements.txt
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
decorator==4.3.0
Django==1.11
django-filter==2.0.0
djangorestframework==3.8.2
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
ipython==5.0.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
libvirt-python==4.4.0
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.13
numpy==1.15.0
paramiko==2.4.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pip2pi==0.7.0
prompt-toolkit==1.0.15
psutil==5.4.6
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
PyNaCl==1.2.1
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.5
requests==2.19.1
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
